:recycle: fix add name to delegate details after refactor, use action not middleware
