+++ src/components/singleTransaction/index.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import { connect } from 'react-redux';
 import { translate } from 'react-i18next';
-import { transactionLoadRequested } from '../../actions/transactions';
+import { loadTransaction } from '../../actions/transactions';
 import TransactionDetails from './../transactions/transactionDetailView';
 import CopyToClipboard from '../copyToClipboard';
 import Box from '../box';
@@ -13,7 +13,7 @@ class SingleTransaction extends React.Component {
   constructor(props) {
     super(props);
     if (props.peers.data) {
-      this.props.transactionLoadRequested({
+      this.props.loadTransaction({
         activePeer: props.peers.data,
         id: this.props.match.params.id,
       });
@@ -23,7 +23,7 @@ class SingleTransaction extends React.Component {
   componentWillReceiveProps(nextProps) {
     if (nextProps.peers.data !== this.props.peers.data
       || nextProps.match.params.id !== this.props.match.params.id) {
-      this.props.transactionLoadRequested({
+      this.props.loadTransaction({
         activePeer: nextProps.peers.data,
         id: nextProps.match.params.id,
       });
@@ -69,7 +69,7 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-  transactionLoadRequested: data => dispatch(transactionLoadRequested(data)),
+  loadTransaction: data => dispatch(loadTransaction(data)),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(translate()(SingleTransaction));
